backend/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12"
anyio==4.9.0 ; python_version >= "3.12"
bidict==0.23.1 ; python_version >= "3.12"
click==8.2.1 ; python_version >= "3.12"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.12"
fastapi-socketio==0.0.10 ; python_version >= "3.12"
fastapi==0.115.12 ; python_version >= "3.12"
h11==0.16.0 ; python_version >= "3.12"
httptools==0.6.4 ; python_version >= "3.12"
idna==3.10 ; python_version >= "3.12"
paho-mqtt==2.1.0 ; python_version >= "3.12"
pydantic-core==2.33.2 ; python_version >= "3.12"
pydantic-settings==2.9.1 ; python_version >= "3.12"
pydantic==2.11.5 ; python_version >= "3.12"
python-dotenv==1.1.0 ; python_version >= "3.12"
python-engineio==4.12.1 ; python_version >= "3.12"
python-socketio==5.13.0 ; python_version >= "3.12"
pyyaml==6.0.2 ; python_version >= "3.12"
simple-websocket==1.1.0 ; python_version >= "3.12"
sniffio==1.3.1 ; python_version >= "3.12"
starlette==0.46.2 ; python_version >= "3.12"
typing-extensions==4.13.2 ; python_version >= "3.12"
typing-inspection==0.4.1 ; python_version >= "3.12"
uvicorn==0.34.2 ; python_version >= "3.12"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.12"
watchfiles==1.0.5 ; python_version >= "3.12"
websockets==15.0.1 ; python_version >= "3.12"
wsproto==1.2.0 ; python_version >= "3.12"
